refactor: 출력 <-> 출력 형식 convert 로직 분리
