+++ src/main/java/racingcar/view/OutputView.java
@@ -2,38 +2,29 @@ package racingcar.view;
 
 import java.util.HashMap;
 import java.util.List;
-import racingcar.controller.domain.Car;
+import racingcar.ListToStringParser;
+import racingcar.domain.Car;
 
 public class OutputView {
+    private final ListToStringParser listToStringParser;
+
+    public OutputView() {
+        listToStringParser = new ListToStringParser();
+    }
 
     public void printRoundOutput(HashMap<Car, Long> carToRacingProgress) {
         StringBuilder sb = new StringBuilder();
         for (Car car : carToRacingProgress.keySet()) {
             sb.append(car.name())
                     .append(" : ")
-                    .append(changeNumberToProgressString(carToRacingProgress.get(car)))
+                    .append(listToStringParser.changeNumberToProgressString(carToRacingProgress.get(car)))
                     .append("\n");
         }
         System.out.println(sb);
     }
 
     public void printFinalWinner(List<String> winnerList) {
-        System.out.println(ConsoleMessage.FINAL_WiNNER + joinListToString(winnerList));
+        System.out.println(ConsoleMessage.FINAL_WiNNER + listToStringParser.joinListToString(winnerList));
     }
 
-    public String joinListToString(List<String> list) {
-        if (list.size() == 1) {
-            return list.getFirst();
-        } else {
-            return String.join(", ", list);
-        }
-    }
-
-    private String changeNumberToProgressString(long progressNumber) {
-        StringBuilder sb = new StringBuilder();
-        for (long i = 0; i < progressNumber; i++) {
-            sb.append("-");
-        }
-        return sb.toString();
-    }
 }
